requirements.txt
fastapi
pydantic
django
django-rest-framework
logging
pytest
requests
uvicorn
sqlalchemy